feat: merge workflows change imports and pathing to match
