+++ INSTALL.md
@@ -1,0 +1,4 @@
+# How to install
+
+submodule
+python virtual env
