+++ docker/NGS_QC/fastq_screen.conf
@@ -35,77 +35,53 @@ DATABASE	Human	/fdb/fastq_screen/FastQ_Screen_Genomes/Human/Homo_sapiens.GRCh38
 ## ftp://ftp.ensembl.org/pub/current/fasta/mus_musculus/dna/
 DATABASE	Mouse	/fdb/fastq_screen/FastQ_Screen_Genomes/Mouse/Mus_musculus.GRCm38
 
-
-
 #########
 ## Rat - sequence available from
 ## ftp://ftp.ensembl.org/pub/current/fasta/rattus_norvegicus/dna/
 DATABASE	Rat	/fdb/fastq_screen/FastQ_Screen_Genomes/Rat/Rnor_6.0 
 
-
-
 ############
 # Drosophila
 DATABASE	Drosophila	/fdb/fastq_screen/FastQ_Screen_Genomes/Drosophila/BDGP6
 
-
-
 #########
 ## Worm
 DATABASE	Worm	/fdb/fastq_screen/FastQ_Screen_Genomes/Worm/Caenorhabditis_elegans.WBcel235
-
-
 
 #########
 ## Yeast - sequence available from
 ## ftp://ftp.ensembl.org/pub/current/fasta/saccharomyces_cerevisiae/dna/
 DATABASE	Yeast	/fdb/fastq_screen/FastQ_Screen_Genomes/Yeast/Saccharomyces_cerevisiae.R64-1-1
 
-
-
 #########
 ## Arabidopsis - sequences available from 
 DATABASE	Arabidopsis	     /fdb/fastq_screen/FastQ_Screen_Genomes/Arabidopsis/Arabidopsis_thaliana.TAIR10
-
-
 
 #########
 ## Ecoli
 ## Sequence available from EMBL accession U00096.2
 DATABASE	Ecoli	/fdb/fastq_screen/FastQ_Screen_Genomes/E_coli/Ecoli
 
-
-
 ##########
 ##rRNA - In house custom database
-DATABASE rRNA /fdb/fastq_screen/FastQ_Screen_Genomes/rRNA/GRCm38_rRNA 
- 
-
+DATABASE    rRNA    /fdb/fastq_screen/FastQ_Screen_Genomes/rRNA/GRCm38_rRNA
 
 ##############
 # Mitochondria
 DATABASE	MT	/fdb/fastq_screen/FastQ_Screen_Genomes/Mitochondria/mitochondria
 
-
-
 ########
 ## PhiX - sequence available from Refseq accession NC_001422.1
 DATABASE	PhiX	     /fdb/fastq_screen/FastQ_Screen_Genomes/PhiX/phi_plus_SNPs
-
-
 
 ##############
 # Lambda
 DATABASE Lambda	/fdb/fastq_screen/FastQ_Screen_Genomes/Lambda/Lambda
 
-
-
 ##########
 ## Vector - Sequence taken from the UniVec database
 ## http://www.ncbi.nlm.nih.gov/VecScreen/UniVec.html
 DATABASE	Vectors	/fdb/fastq_screen/FastQ_Screen_Genomes/Vectors/Vectors
-
-
 
 ############
 ## Adapters - sequence derived from the FastQC contaminats file
