--- workflow/illumina_demux/Snakefile
@@ -1,54 +1,0 @@
-
-import os
-
-
-envvars: "SNK_CONFIG"
-configfile: os.environ['SNK_CONFIG']
-
-
-demux_runs_projects = {
-    'run': config['runs'],
-    'project': config['projects'],
-    'out_to': config['out_to'],
-    'run_ids': config['run_ids'],
-    'rnums': config['rnums'],
-    'reads_out': config['reads_out']
-}
-
-
-rule all:
-    input:
-        # ~~ illumnia demultiplexing ~~
-        expand("{run}", **demux_runs_projects),
-        expand("{out_to}/{project}/{reads_out}", **demux_runs_projects),
-        expand("{out_to}/Stats/Stats.json", **demux_runs_projects),
-        expand("{out_to}/.DEMUX_COMPLETE", **demux_runs_projects)
-        
-
-rule bcl2fastq:
-    """
-        BCL to FASTQ file converter
-        bcl2fastq v2.20.0.422
-    """
-    input:
-        run_dir                = expand("{run}", **demux_runs_projects),
-        binary_base_calls      = expand("{files}", files=config['bcl_files']),
-        samplesheets           = expand("{run}/SampleSheet.csv", **demux_runs_projects)
-    output:
-        seq_data               = expand("{out_to}/{project}/{reads_out}", **demux_runs_projects),
-        undetermined           = expand("{out_to}/Undetermined_S0_R{rnums}_001.fastq.gz", **demux_runs_projects),
-        stats                  = expand("{out_to}/Stats/Stats.json", **demux_runs_projects),
-        breadcrumb             = expand("{out_to}/.DEMUX_COMPLETE", **demux_runs_projects),
-    params:
-        out_dir                = expand("{out_to_directory}", out_to_directory=config['out_to']),
-    containerized: config["resources"]["sif"] + "bcl2fastq.sif"
-    threads: 24
-    shell: 
-        r"""
-            bcl2fastq \
-            --runfolder-dir {input.run_dir} \
-            --min-log-level=INFO \
-            -r {threads} -p {threads} -w {threads} \
-            --no-lane-splitting -o {params.out_dir}
-            touch {params.out_dir}/.DEMUX_COMPLETE
-        """
