--- workflow/ngs_qaqc/scripts
@@ -1,1 +1,0 @@
-../../scripts
